+++ client/app/expense/components/ExpenseForm.tsx
@@ -50,11 +50,11 @@ const ExpenseForm: React.FC<ExpenseFormProps> = ({
 	};
 
 	return (
-		<Card>
+		<Card className="min-w-fit max-h-fit">
 			<CardContent className="flex flex-col gap-2 p-3">
 				<h1 className="font-bold">{"Ajouter ma dépense"}</h1>
 				<form onSubmit={handleSubmit} className="flex gap-2">
-					<div className="w-full flex flex-col md:flex-row gap-2">
+					<div className="w-full flex flex-col gap-2">
 						<Input
 							type="text"
 							placeholder="Dépense"
